HMW-47 Moved the Virtualized list into a separate component for easier reuse and converted the virtualized list on the state page into using this new component.
